23597: Move RefineYourSearch to containers. Add it to Location SearchResults
